Correction des classes batiments et adaptation d'Ouvrier
