+++ src/batiments/TypeBatiment.java
@@ -3,15 +3,20 @@ package batiments;
 public enum TypeBatiment {
 
     // Lieux de ressources
+	LIEU_DE_RESSOURCE, // valeur provisoire, on n'a pas implémenté les différents types de lieux de ressources pour l'instant
     FORET,
     MINE,
     CARRIERE,
     PUITS_PETROLE,
 
     // Usines
+	USINE, //valeur provisoire, on n'a pas implémenté les différents types d'usines pour l'instant
     MENUISERIE,
     FONDERIE,
     RAFFINERIE,
     USINE_ELECTRONIQUE,
-    ZONE_ASSEMBLAGE
+    ZONE_ASSEMBLAGE,
+
+	// Bâtiments de soutien
+	LIEUX_DE_REPOS,
 }
