+++ src/entites/Ouvrier.java
@@ -35,19 +35,6 @@ public class Ouvrier implements Item {
         }
     }
 
-    public enum EtatOuvrier {
-        TRES_FATIGUE (0.3),
-        FATIGUE      (0.8),
-        NORMAL       (1.0),
-        MOTIVE       (1.2),
-        TRES_MOTIVE  (1.7);
-
-        public final double multiplicateur;
-
-        EtatOuvrier(double multiplicateur) {
-            this.multiplicateur = multiplicateur;
-        }
-    }
 
     private String nom;
     private int x, y;
